Move api stuff into one place
