+++ app/src/main/java/com/khatm/client/factories/ApiFactory.kt
@@ -40,5 +40,36 @@ object ApiFactory {
         .build()
 
 
-    val khatmApi : KhatmApi = retrofit.create(KhatmApi::class.java)
+    val api : KhatmApi = retrofit.create(KhatmApi::class.java)
+
+    suspend fun <T : Any> call(call: suspend () -> Response<T>, errorMessage: String): T? {
+
+        val result : Result<T> = safeApiResult(call, errorMessage)
+        var data : T? = null
+
+        when(result) {
+            is Result.Success ->
+                data = result.data
+            is Result.Error -> {
+                // TODO: If 404 then log user out
+                Log.e("RepositoryBase", "$errorMessage; ${result.exception}")
+            }
+        }
+
+        return data
+    }
+
+    private suspend fun <T: Any> safeApiResult(call: suspend ()-> Response<T>, errorMessage: String) : Result<T> {
+        val response = call.invoke()
+        if (response.isSuccessful) {
+            return Result.Success(response.body()!!)
+        }
+
+        return Result.Error(IOException("API Error - $errorMessage"))
+    }
+}
+
+sealed class Result<out T: Any> {
+    data class Success<out T : Any>(val data: T) : Result<T>()
+    data class Error(val exception: Exception) : Result<Nothing>()
 }
